combined test_sierra_adapters into test_adapters
